backend/requirements.txt
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
openai>=1.0.0
PyYAML>=6.0.1
jsonschema>=4.22.0
